<BRMS-2852> Split the Modal and Component definitions for Import.
