+++ modules/editor/src/index.js
@@ -2,5 +2,6 @@
 
 export { Toolbox } from './toolbox.js';
 export { ImportModal } from './import-modal.js';
+export { ImportComponent } from './import-component.js';
 export { ImportData, ValidImportData, InvalidImportData, parseImport } from './lib/importer.js';
 export { ExportParameters, toKml, toWkt, toGeoJson, toStats, toIds } from './lib/exporter.js';
